[DROOLS-5897] Compound BigDecimal operator is translated to assignOperator in MVEL (#3318)
- This test case should pass once Drools consumes Mvel 2.4.12+
