+++ drools-model/drools-model-compiler/src/test/java/org/drools/modelcompiler/MvelDialectTest.java
@@ -743,4 +743,46 @@ public class MvelDialectTest extends BaseModelTest {
         Assertions.assertThat(names).containsExactlyInAnyOrder("Mario", "Luca", "Leonardo");
         Assertions.assertThat(addresses).contains("Milan");
     }
+
+    @Test
+    public void testCompoundOperator() throws Exception {
+        // DROOLS-5894 // DROOLS-5901 // DROOLS-5897
+        String drl =
+                "import " + Person.class.getCanonicalName() + "\n" +
+                "import " + BigDecimal.class.getCanonicalName() + "\n" +
+                "dialect \"mvel\"\n" +
+                "rule R\n" +
+                "when\n" +
+                "    $p : Person( age >= 26 )\n" +
+                "then\n" +
+                "    BigDecimal result = 0B;" +
+                "    $p.money += 50000;\n" + // 50000
+                "    $p.money -= 10000;\n" + // 40000
+                "    $p.money /= 10;\n" + // 4000
+                "    $p.money *= 10;\n" + // 40000
+                "    $p.money += $p.money;\n" + // 80000
+                "    $p.money /= $p.money;\n" + // 1
+                "    $p.money *= $p.money;\n" + // 1
+                "    $p.money -= $p.money;\n" + // 0
+                "    BigDecimal anotherVar = 10B;" +
+                "    $p.money += anotherVar;\n" + // 10
+                "    $p.money /= anotherVar;\n" + // 1
+                "    $p.money *= anotherVar;\n" + // 1
+                "    $p.money -= anotherVar;\n" + // 0
+                "    int intVar = 20;" +
+                "    $p.money += intVar;\n" + // 20
+                "    $p.money /= intVar;\n" + // 1
+                "    $p.money *= intVar;\n" + // 1
+                "    $p.money -= intVar;\n" + // 0
+                "end";
+
+        KieSession ksession = getKieSession(drl);
+
+        Person john = new Person("John", 30);
+        john.setMoney( new BigDecimal( 70000 ) );
+
+        ksession.insert(john);
+        assertEquals(1, ksession.fireAllRules());
+        assertEquals(new BigDecimal( 0 ), john.getMoney());
+    }
 }
